chore: rename setupTwig file to utils
